feat: move shorturl to its own domain
