+++ internal/domain/shorturl/main_test.go
@@ -1,4 +1,4 @@
-package shortener
+package shorturl
 
 import (
 	"crypto/md5"
